Projet OPA/Data et Code/Partie 3/requirements.txt
fastapi==0.104.1
pandas==2.0.3
psycopg2_binary==2.9.7
python-decouple==3.8
python_dateutil==2.8.2
Requests==2.31.0
scikit_learn==1.3.2
SQLAlchemy==2.0.23
tqdm==4.66.1
pydantic==2.5.2
uvicorn==0.24.0